refactor:[2-10] SessionImage 변수들 클래스 분리
